backend/requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
sqlparse==0.4.4
typing_extensions==4.10.0
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.10.0
